+++ asreview/webapp/src/api/index.js
@@ -1,2 +1,3 @@
+export { default as axiosErrorHandler } from "./axiosErrorHandler";
 export { default as BaseAPI } from "./BaseAPI";
 export { default as ProjectAPI } from "./ProjectAPI";
